requirements.txt
click==8.1.3
certifi==2023.7.22
dnspython==2.2.1
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.3.3
python-dotenv==0.21.0
Werkzeug==2.2.2